+++ @remirror/core/src/extension-manager.helpers.ts
@@ -107,6 +107,45 @@ export const createCommands = ({ extensions, params }: CreateCommandsParams) => 
 
   return items;
 };
+interface CreateHelpersParams extends ExtensionListParams {
+  /**
+   * The params which are passed to each extensions `helpers` method.
+   */
+  params: ExtensionManagerParams;
+}
+
+/**
+ * Generate all the helpers from the extension list.
+ *
+ * Helpers are functions which enable extensions to provide useful
+ * information or transformations to their consumers and other extensions.
+ */
+export const createHelpers = ({ extensions, params }: CreateHelpersParams) => {
+  const getItemParams = (extension: Required<Pick<AnyExtension, 'helpers'>>) =>
+    extension.helpers({
+      ...params,
+      ...(isMarkExtension(extension)
+        ? { type: params.schema.marks[extension.name] }
+        : isNodeExtension(extension)
+        ? { type: params.schema.nodes[extension.name] }
+        : ({} as any)),
+    });
+
+  const items: Record<string, AnyFunction> = {};
+  const names = new Set<string>();
+
+  extensions.filter(hasExtensionProperty('helpers')).forEach(currentExtension => {
+    const item = getItemParams(currentExtension);
+
+    Object.entries(item).forEach(([name, helper]) => {
+      isNameUnique({ name, set: names, shouldThrow: true });
+
+      items[name] = helper;
+    });
+  });
+
+  return items;
+};
 
 /**
  * Checks to see if an optional property exists on an extension.