+++ changelog.md
@@ -11,6 +11,7 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 - 🚀 `@remirror/dev`: New package for aiding in developing new extensions and components.
 - 🚀 `@remirror/core`: Introduce the concept of meta tags for extensions. These allow an extension to tag itself and these tags are made available through the tag object which is passed to all extension methods.
+- 🚀 `@remirror/core`: Add a `helpers` method to extensions. These are similar to commands except they don't have access to the view and shouldn't directly affect the editor. They can also return data and receive custom parameters. They can be accessed with `manager.data.helpers.myHelper()`.
 - 🚀 `@remirror/core-extensions`: Add `TrailingNodeExtension` to always append a specified node to the end of the dom.
 - `@remirror/core`: Add `getExtraAttrs` method to the extension which can be used in the `(Mark/Node)Extension`.
 - 🚀 `@remirror/core`: Add `DropCursorExtension` for a cursor to show up at the exact location an item will be dropped.